Reorganize data publication code into `publish` module
